plot preparations moved from QtGUI to rantest.py
