+++ dcstats/rantest.py
@@ -84,16 +84,17 @@ class RantestBinomial():
 
 
 class RantestContinuous():
-    def __init__(self, X, Y, are_paired=False):
+    def __init__(self, df, are_paired=False):
         """ 
         Parameters
         ----------
-        X : observations in first trial, list of floats
-        Y : observations in second trial, list of floats
+        df : pandas DataFrame with two columns containing two samples 
         are_paired : are observations paired, boolean
         """   
-        self.X, self.Y = X, Y
-        self.nx, self.ny = len(X), len(Y)
+        self.df = df
+        self.X = df.iloc[:,0].dropna().values #.tolist()
+        self.Y = df.iloc[:,1].dropna().values #.tolist()
+        self.nx, self.ny = len(self.X), len(self.Y)
         self.are_paired = are_paired
         random.seed(1984)
         np.random.seed(1984)
@@ -112,7 +113,6 @@ class RantestContinuous():
         #option to have bootstrap calculated CIs should go here
         return str(ttc) + str(hedges_calculation)
 
-            
     def run_rantest(self, nran):
         """ Resample without replacement nran times and get statistics of obtained distribution. """
         self.nran = nran
@@ -137,6 +137,43 @@ class RantestContinuous():
         self.pequal = self.nequal / float(self.nran)
         self.lo95lim = np.percentile(self.randiff, 2.5)
         self.hi95lim = np.percentile(self.randiff, 97.5)
+
+    
+
+
+    #    self.ax1 = self.figure.add_subplot(1, 2, 1)
+    #    self.ax2 = self.figure.add_subplot(1, 2, 2)
+
+    def add_boxplot(self, X, Y, sample_names=None):
+        if sample_names is None:
+            names = ['sample 1', 'sample 2']
+        else:
+            names = sample_names
+        self.ax1.clear()
+        self.ax1.boxplot((X, Y))
+        # Add some random "jitter" to the x-axis
+        x = np.random.normal(1, 0.04, size=len(X))
+        self.ax1.plot(x, X, '.', alpha=0.4)
+        y = np.random.normal(2, 0.04, size=len(Y))
+        self.ax1.plot(y, Y, '.', alpha=0.4)
+        plt.setp(self.ax1, xticks=[1, 2], xticklabels=names)
+        self.ax1.set_ylabel('measurment values')
+        
+        
+
+    def add_randhisto(self, randiff, dbar, lo95lim, hi95lim):
+        self.ax2.clear()
+        self.ax2.hist(randiff, bins=20)
+        self.ax2.axvline(x=dbar, color='r', label='observed difference')
+        self.ax2.axvline(x=-dbar, color='r')
+        self.ax2.axvline(x=lo95lim, color='k', linestyle='--', label='2.5% limits')
+        self.ax2.axvline(x=hi95lim, color='k', linestyle='--')
+        self.ax2.set_xlabel('difference between means')
+        self.ax2.set_ylabel('frequency')
+        self.ax2.legend(bbox_to_anchor=(0., 1.02, 1., .102), loc=3,
+                        borderaxespad=0.)
+        plt.tight_layout()
+
         
     def __repr__(self):
         return ('\nRantest:  {0:d} randomisations'.format(self.nran) +
@@ -161,20 +198,22 @@ class RantestBatch():
     def run_rantest(self, nran):
         for i in range(self.n-1):
             for j in range(i+1, self.n):
-                rnt = RantestContinuous(self.df.iloc[:, i].dropna().values.tolist(), 
-                                        self.df.iloc[:, j].dropna().values.tolist())
+                #rnt = RantestContinuous(self.df.iloc[:, i].dropna().values, 
+                #                        self.df.iloc[:, j].dropna().values)
+                rnt = RantestContinuous(self.df.iloc[:, [i, j]])
                 rnt.run_rantest(nran)
                 self.log.append('\n*****   ' + self.names[i] + ' versus ' + 
                                 self.names[j] + '   *****')
                 self.log.append(rnt.describe_data())
                 self.log.append(str(rnt))
 
-    def get_boxplot(self):
-        figure = plt.figure()
-        ax = figure.add_subplot(1, 1, 1)
-        ax = self.df.boxplot()
-        for i in range(self.df.shape[1]):
-            X = self.df.iloc[:, i].dropna().values #.tolist()
-            x = np.random.normal(i+1, 0.04, size=len(X))
-            ax.plot(x, X, '.', alpha=0.4)
-        return figure
+
+def get_boxplot(df):
+    figure = plt.figure()
+    ax = figure.add_subplot(1, 1, 1)
+    ax = df.boxplot()
+    for i in range(df.shape[1]):
+        X = df.iloc[:, i].dropna().values #.tolist()
+        x = np.random.normal(i+1, 0.04, size=len(X))
+        ax.plot(x, X, '.', alpha=0.4)
+    return figure
