+++ tests/test_rantest.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
+import pandas
+
 from dcstats.rantest import RantestBinomial
 from dcstats.rantest import RantestContinuous
 #from test_statistics import isclose
@@ -9,11 +11,12 @@ def test_regression_rantest_continuos():
     # Samples from treatment T1 and T2
     T1 = [100, 108, 119, 127, 132, 135, 136] #, 164]
     T2 = [122, 130, 138, 142, 152, 154, 176]
-    rnt = RantestContinuous(T1, T2, True)    
+    df = pandas.DataFrame({'Sample1':T1, 'Sample2':T2})
+    rnt = RantestContinuous(df, True)    
     rnt.run_rantest(5000)
     assert (rnt.p2tail == 0.018)
     assert (rnt.n2tail == 90)
-    rnt = RantestContinuous(T1, T2, False)
+    rnt = RantestContinuous(df, False)
     rnt.run_rantest(5000)    
     assert (rnt.p2tail == 0.0178)
     assert (rnt.n2tail == 89)
